libs(layout:feature): moved logic in shared layout lib to feature
