+++ src/spas/libs/layout/feature/src/lib/toolbar/toolbar-default/toolbar-default.component.html
@@ -1,1 +1,32 @@
-<p>toolbar-default works!</p>
+<!-- toolbar -->
+<div class="toolbar">
+    <!-- mobile side nav toggle -->
+
+    <!-- logo -->
+
+    <!-- search bar-->
+
+    <!-- full screen toggle -->
+
+    <!-- search -->
+
+    <!-- notifications -->
+    <ui-toolbar-notifications></ui-toolbar-notifications>
+    <!-- /notifications -->
+
+    <!-- toolbar user profile -->
+    <ui-toolbar-user class="toolbar-btn"></ui-toolbar-user>
+    <!-- /toolbar user profile -->
+
+    <!-- quick panel -->
+</div>
+<!-- /toolbar -->
+
+<p>hello!!!</p>
+
+<!-- <p>
+    is navigation enabled:{{isNavigationEnabled}}
+</p>
+<p>
+    is top navigation: {{isTopNavigation}}
+</p> -->
