+++ src/spas/libs/shared/ui/actions/src/lib/clicked-outside/clicked-outside.module.ts
@@ -1,8 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+
 import { ClickedOutsideDirective } from './clicked-outside.directive';
-
-
 
 @NgModule({
   declarations: [ClickedOutsideDirective],
