+++ src/spas/libs/shared/ui/layout/src/lib/shared-ui-layout.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { ToolbarModule } from './toolbar/toolbar.module';
+import { ToolbarModule } from './toolbar/shared-ui-toolbar.module';
 import { ToolbarComponent } from './toolbar/toolbar.component';
 import { ToolbarUserComponent } from './toolbar/toolbar-user/toolbar-user.component';
-import { SharedUiActionsModule } from '@nimb3s/shared/ui/actions';
+import { ToolbarNotificationsComponent } from './toolbar/toolbar-notifications/toolbar-notifications.component';
 
 @NgModule({
   imports: [
@@ -12,7 +12,9 @@ import { SharedUiActionsModule } from '@nimb3s/shared/ui/actions';
     ToolbarModule,
   ],
   exports: [
-    ToolbarComponent
+    ToolbarComponent,
+    ToolbarUserComponent,
+    ToolbarNotificationsComponent
   ]
 })
 export class SharedUiLayoutModule {}
