+++ src/spas/libs/shared/ui/layout/src/lib/toolbar/shared-ui-toolbar.module.ts
@@ -1,10 +1,11 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
-import { ToolbarComponent } from '../toolbar/toolbar.component';
-import { ToolbarUserComponent } from './toolbar-user/toolbar-user.component';
 import { SharedUiActionsModule } from '@nimb3s/shared/ui/actions';
-import { ToolbarNotificationsComponent } from '../toolbar/toolbar-notifications/toolbar-notifications.component';
+
+import { ToolbarComponent } from './toolbar.component';
+import { ToolbarUserComponent } from './toolbar-user/toolbar-user.component';
+import { ToolbarNotificationsComponent } from './toolbar-notifications/toolbar-notifications.component';
 
 @NgModule({
   declarations: [
@@ -19,6 +20,7 @@ import { ToolbarNotificationsComponent } from '../toolbar/toolbar-notifications/
   ],
   exports: [
     ToolbarComponent,
+    ToolbarUserComponent,
     ToolbarNotificationsComponent
   ]
 })
